# Optimize request latency