--- a/types/exchangeTxs.go
+++ b/types/exchangeTxs.go
@@ -1,5 +1,11 @@
 package types
 
+import (
+	"bytes"
+	"crypto/sha256"
+	"encoding/binary"
+)
+
 type PoolTxEvent struct {
 	EthTxHeader
 	PoolLocation
@@ -31,4 +37,28 @@ type PoolRangeFields struct {
 	AskTick   int  `json:"askTick"`
 	IsBuy     bool `json:"isBuy"`
 	InBaseQty bool `json:"inBaseQty"`
+}
+
+func (p PoolTxEvent) Hash() [32]byte {
+	buf := new(bytes.Buffer)
+	buf.Grow(270)
+	binary.Write(buf, binary.BigEndian, int32(p.BlockNum))
+	buf.WriteString(string(p.TxHash))
+	binary.Write(buf, binary.BigEndian, int32(p.TxTime))
+	buf.WriteString(string(p.User))
+
+	buf.WriteString(string(p.ChainId))
+	buf.WriteString(string(p.Base))
+	buf.WriteString(string(p.Quote))
+	binary.Write(buf, binary.BigEndian, int32(p.PoolIdx))
+
+	buf.WriteString(string(p.EntityType))
+	buf.WriteString(string(p.ChangeType))
+	buf.WriteString(string(p.PositionType))
+
+	binary.Write(buf, binary.BigEndian, int32(p.BidTick))
+	binary.Write(buf, binary.BigEndian, int32(p.AskTick))
+	binary.Write(buf, binary.BigEndian, p.IsBuy)
+	binary.Write(buf, binary.BigEndian, p.InBaseQty)
+	return sha256.Sum256(buf.Bytes())
 }