Firstly, this PR removes dependency on `structhash` for generating IDs for positions, limit orders, and transactions. Instead a buffer is manually filled with all the relevant fields and hashed. This noticeably speeds up requests with a lot of these IDs, mostly `pool_positions` (down to ~30ms from ~90ms).

But the slowest request by far is `pool_txs` because it retrieves all transactions and then sorts them before returning a handful of TXs which were sent most recently. But the transactions get inserted into memory cache in an almost sorted order already, and they're never sorted in any other way. 

So the second change of this PR is to maintain the order during insertion, and retrieve the last rows directly. This brings the latency of `pool_txs` request down to less than 1ms compared to 150-300ms currently.

Future optimizations:
1. The same ordering optimization should be applied to pool positions, but they're currently stored in a map and also they'll require re-sorting on every update since they're sorted by last update time. Not as trivial, but inevitable as the state will keep growing.
2. `pool_txs` with `time` parameter still takes longer than it reasonably should (200-300ms), but not 5 seconds like it used to. Though it's very rarely used.